Re-code count section
