+++ app/src/main/AndroidManifest.xml
@@ -21,7 +21,7 @@
 
         </activity>
 
-        <activity android:name=".UI.BasicCounterActivity">
+        <activity android:name=".UI.CounterActivity">
 
         </activity>
