+++ app/src/main/java/com/example/appraisal/backend/CountTrial.java
@@ -1,4 +1,17 @@
 package com.example.appraisal.backend;
 
 public class CountTrial extends Trial {
+    private int counter;
+
+    public CountTrial() {
+        counter = 0;
+    }
+
+    public void increase() {
+        counter++;
+    }
+
+    public int getCount() {
+        return counter;
+    }
 }
