+++ app/src/main/java/com/example/appraisal/model/CounterModel.java
@@ -1,0 +1,2 @@
+package com.example.appraisal.model;public class CounterModel {
+}
